requirements.txt
streamlit
streamlit-javascript
langgraph
langchain_community
tiktoken
pinecone
langchain-openai
langchain_pinecone
langchainhub
chromadb
langchain
tavily-python
python-dotenv
bs4
starlette
itsdangerous
uvicorn
fastapi
sqlalchemy
psycopg2-binary
requests
anthropic
langchain-anthropic